2008-01-09  haytjes  <hv1989@gmail.com>
	* src/awn-bar.c:
	Fix #175210: Cairo drawing bugs

+++ src/awn-bar.c
@@ -314,7 +314,7 @@ render (AwnBar *bar, cairo_t *cr, gint x_width, gint height)
 	cairo_set_operator (cr, CAIRO_OPERATOR_SOURCE);
 	cairo_paint (cr);
 	
-	double x = (settings->monitor.width-width)*settings->bar_pos;
+	double x = (int)((settings->monitor.width-width)*settings->bar_pos);
 	
 	if (settings->bar_angle != 0)
         {
